refactor(CustomJwtAuthenticationEntryPoint) : JWT 인증 실패 예외 처리 방식 변경
기존 response 설정 방식 대신 예외를 던지는 방식으로 변경하여 예외 처리를 일관되게 유지.
JwtException을 활용해 JWT 관련 예외를 명확하게 구분할 수 있도록 수정.
