--- src/main/java/org/terning/terningserver/auth/jwt/CustomJwtAuthenticationEntryPoint.java
@@ -1,39 +1,0 @@
-package org.terning.terningserver.auth.jwt;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-import org.terning.terningserver.exception.CustomException;
-
-import java.io.IOException;
-
-import static jakarta.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static org.terning.terningserver.exception.enums.ErrorMessage.INVALID_TOKEN;
-
-@Component
-@RequiredArgsConstructor
-public class CustomJwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            AuthenticationException exception
-    ) throws IOException {
-        setResponse(response);
-    }
-
-    private void setResponse(HttpServletResponse response) throws IOException {
-        response.setCharacterEncoding("UTF-8");
-        response.setContentType(APPLICATION_JSON_VALUE);
-        response.setStatus(SC_UNAUTHORIZED);
-        response.getWriter().println(objectMapper.writeValueAsString(new CustomException(INVALID_TOKEN)));
-    }
-}
